Alterações feitas para integração com o BERTimbau.

+++ database/migrations/2024_12_12_143433_remove_is_admin_from_usuarios_table.php
@@ -4,23 +4,21 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
+class RemoveIsAdminFromUsuariosTable extends Migration
 {
-    /**
-     * Run the migrations.
-     */
-    public function up(): void
-{
-    Schema::table('usuarios', function (Blueprint $table) {
-        $table->dropColumn('is_admin');
-    });
-}
+    public function up()
+    {
+        if (Schema::hasColumn('usuarios', 'is_admin')) {
+            Schema::table('usuarios', function (Blueprint $table) {
+                $table->dropColumn('is_admin');
+            });
+        }
+    }
 
-public function down(): void
-{
-    Schema::table('usuarios', function (Blueprint $table) {
-        $table->boolean('is_admin')->default(false);
-    });
+    public function down()
+    {
+        Schema::table('usuarios', function (Blueprint $table) {
+            $table->boolean('is_admin')->default(false);
+        });
+    }
 }
-
-};
